# src/main/resources/docs/api-docs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8081/atm-monitoring",
      "description": "Generated server url"
    }
  ],
  "paths": {
    "/v1/generate-token": {
      "post": {
        "tags": [
          "token-generation-for-testing-controller"
        ],
        "operationId": "generateAccessToken",
        "parameters": [
          {
            "name": "empCode",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/v1/video-logs/atm/{atmId}": {
      "get": {
        "tags": [
          "video-log-controller"
        ],
        "summary": "Get video logs by time range",
        "operationId": "findByTimeRange",
        "parameters": [
          {
            "name": "startTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Access is denied. User does not have the 'ROLE_OWNER' role",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoLog"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Video logs retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoLog"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoLog"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/video-logs/atm/{atmId}/transaction/{transactionId}": {
      "get": {
        "tags": [
          "video-log-controller"
        ],
        "summary": "Get video logs by ATM ID and transaction ID",
        "operationId": "getVideoLogsByAtmIdAndTransactionId",
        "parameters": [
          {
            "name": "atmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "transactionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Access is denied. User does not have the 'ROLE_OWNER' role",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoLog"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Video logs retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoLog"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoLog"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/video-logs/atm/{atmId}/customer/{customerId}": {
      "get": {
        "tags": [
          "video-log-controller"
        ],
        "summary": "Get video logs by ATM ID, customer ID, and time range",
        "operationId": "getVideoLogsByAtmIdAndCustomerIdAndTimeRange",
        "parameters": [
          {
            "name": "atmId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endTime",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Access is denied. User does not have the 'ROLE_OWNER' role",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoLog"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Video logs retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoLog"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VideoLog"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Get a list of transactions filtered by transaction type",
        "operationId": "getTransactionByType",
        "parameters": [
          {
            "name": "transactionType",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "403": {
            "description": "Access is denied. User does not have the 'ROLE_OWNER' role",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid transaction type provided",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Transactions retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Transaction"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/transactions/customers/count/24-hours": {
      "get": {
        "tags": [
          "transaction-controller"
        ],
        "summary": "Get the total count of customers who made transactions in the last 24 hours",
        "operationId": "getTotalTransactionsCount",
        "responses": {
          "403": {
            "description": "Access is denied. User does not have the 'ROLE_OWNER' role",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          },
          "200": {
            "description": "Total count retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "integer",
                  "format": "int32"
                }
              }
            }
          }
        }
      }
    },
    "/v1/failures": {
      "get": {
        "tags": [
          "failure-log-controller"
        ],
        "summary": "Get all failure logs with their causes of failure",
        "operationId": "getFailures",
        "responses": {
          "403": {
            "description": "Access is denied. User does not have the 'ROLE_OWNER' role",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FailureLog"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FailureLog"
                  }
                }
              }
            }
          },
          "200": {
            "description": "Failure logs retrieved successfully",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FailureLog"
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "VideoLog": {
        "type": "object",
        "properties": {
          "transactionId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "videoUrl": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Transaction": {
        "type": "object",
        "properties": {
          "transactionId": {
            "type": "string"
          },
          "customerId": {
            "type": "string"
          },
          "transactionType": {
            "type": "string",
            "enum": [
              "CASH_WITHDRAWAL",
              "DEPOSIT",
              "BALANCE_INFORMATION"
            ]
          },
          "amount": {
            "type": "number",
            "format": "double"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "FailureLog": {
        "type": "object",
        "properties": {
          "atmId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "failureType": {
            "type": "string",
            "enum": [
              "CASH_DISPENSER_ERROR",
              "NETWORK_TIMEOUT",
              "CARD_READER_FAILURE",
              "PRINTER_ERROR",
              "SYSTEM_CRASH",
              "UNKNOWN_ERROR"
            ]
          },
          "errorCode": {
            "type": "string"
          },
          "transactionId": {
            "type": "string"
          }
        }
      }
    }
  }
}